+++ Qwerty.ECS.Runtime/Archetypes/EcsArchetypeManager.cs
@@ -12,26 +12,26 @@ namespace Qwerty.ECS.Runtime.Archetypes
         
         private readonly EcsArchetype m_emptyArchetype;
         private readonly List<EcsArchetype> m_archetypes = new List<EcsArchetype>();
-        private readonly byte[] m_indicesBuffer;
+        private readonly short[] m_indicesBuffer;
         private readonly EcsWorldSetting m_setting;
 
         public EcsArchetypeManager(EcsWorldSetting setting)
         {
             m_setting = setting;
-            m_indicesBuffer = new byte[EcsTypeManager.typeCount];
-            m_emptyArchetype = new EcsArchetype(m_archetypes.Count, Array.Empty<byte>(), m_setting);
+            m_indicesBuffer = new short[EcsTypeManager.typeCount];
+            m_emptyArchetype = new EcsArchetype(m_archetypes.Count, Array.Empty<short>(), m_setting);
             m_archetypes.Add(m_emptyArchetype);
         }
         
-        internal EcsArchetype FindOrCreateArchetype(byte[] indicesBuffer, int indicesLen)
+        internal EcsArchetype FindOrCreateArchetype(short[] indicesBuffer, int indicesLen)
         {
             EcsArchetype current = m_emptyArchetype;
             for (int i = 0; i < indicesLen; i++)
             {
-                byte index = indicesBuffer[i];
+                short index = indicesBuffer[i];
                 if (!current.next.TryGetValue(index, out int nextId))
                 {
-                    byte[] newIndices = new byte[i + 1];
+                    short[] newIndices = new short[i + 1];
                     Array.Copy(indicesBuffer, newIndices, i + 1);
                     EcsArchetype next = new EcsArchetype(m_archetypes.Count, newIndices, m_setting);
                     m_archetypes.Add(next);
@@ -44,7 +44,7 @@ namespace Qwerty.ECS.Runtime.Archetypes
             return current;
         }
 
-        internal EcsArchetype FindOrCreateNextArchetype(EcsArchetype archetype, byte nextIndex)
+        internal EcsArchetype FindOrCreateNextArchetype(EcsArchetype archetype, short nextIndex)
         {
             if (archetype.next.TryGetValue(nextIndex, out int next))
             {
@@ -54,8 +54,8 @@ namespace Qwerty.ECS.Runtime.Archetypes
             bool added = false;
             int len = 0;
             
-            byte[] typeIndices = archetype.indices;
-            foreach (byte typeIndex in typeIndices)
+            short[] typeIndices = archetype.indices;
+            foreach (short typeIndex in typeIndices)
             {
                 if (nextIndex < typeIndex && !added)
                 {
@@ -72,7 +72,7 @@ namespace Qwerty.ECS.Runtime.Archetypes
             return FindOrCreateArchetype(m_indicesBuffer, len);
         }
         
-        internal EcsArchetype FindOrCreatePriorArchetype(EcsArchetype archetype, byte priorIndex)
+        internal EcsArchetype FindOrCreatePriorArchetype(EcsArchetype archetype, short priorIndex)
         {
             if (archetype.prior.TryGetValue(priorIndex, out int prior))
             {
@@ -80,8 +80,8 @@ namespace Qwerty.ECS.Runtime.Archetypes
             }
             
             int len = 0;
-            byte[] typeIndices = archetype.indices;
-            foreach (byte typeIndex in typeIndices)
+            short[] typeIndices = archetype.indices;
+            foreach (short typeIndex in typeIndices)
             {
                 if (typeIndex != priorIndex)
                 {
