+++ Qwerty.ECS.Tests/EcsWorldTest.Filter.cs
@@ -59,40 +59,40 @@ namespace Qwerty.ECS.Tests
         [Test]
         public void FilterAllTest()
         {
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().AllOf<ComponentA>().all).Count());
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentA>().all).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().AllOf<ComponentA>().all).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentA>().all).Count());
             
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB>().all).Count());
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentB>().all).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB>().all).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentB>().all).Count());
             
-            Assert.AreEqual(3, new HashSet<byte> {0, 1, 2}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC>().all).Count());
-            Assert.AreEqual(3, new HashSet<byte> {0, 1, 2}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentC>().all).Count());
+            Assert.AreEqual(3, new HashSet<short> {0, 1, 2}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC>().all).Count());
+            Assert.AreEqual(3, new HashSet<short> {0, 1, 2}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentC>().all).Count());
             
-            Assert.AreEqual(4, new HashSet<byte> {0, 1, 2, 3}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD>().all).Count());
-            Assert.AreEqual(4, new HashSet<byte> {0, 1, 2, 3}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentD>().all).Count());
+            Assert.AreEqual(4, new HashSet<short> {0, 1, 2, 3}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD>().all).Count());
+            Assert.AreEqual(4, new HashSet<short> {0, 1, 2, 3}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentD>().all).Count());
             
-            Assert.AreEqual(5, new HashSet<byte> {0, 1, 2, 3, 4}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentE>().all).Count());
-            Assert.AreEqual(5, new HashSet<byte> {0, 1, 2, 3, 4}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentE, ComponentE>().all).Count());
+            Assert.AreEqual(5, new HashSet<short> {0, 1, 2, 3, 4}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentE>().all).Count());
+            Assert.AreEqual(5, new HashSet<short> {0, 1, 2, 3, 4}.Intersect(new EcsFilter().AllOf<ComponentA, ComponentB, ComponentC, ComponentD, ComponentE, ComponentE>().all).Count());
         }
         
         [Test]
         public void FilterAnyTest()
         {
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().AnyOf<ComponentA>().any).Count());
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentA>().any).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().AnyOf<ComponentA>().any).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentA>().any).Count());
             
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentB>().any).Count());
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentB, ComponentB>().any).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentB>().any).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().AnyOf<ComponentA, ComponentB, ComponentB>().any).Count());
         }
         
         [Test]
         public void FilterNoneTest()
         {
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().NoneOf<ComponentA>().none).Count());
-            Assert.AreEqual(1, new HashSet<byte> {0}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentA>().none).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().NoneOf<ComponentA>().none).Count());
+            Assert.AreEqual(1, new HashSet<short> {0}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentA>().none).Count());
             
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentB>().none).Count());
-            Assert.AreEqual(2, new HashSet<byte> {0, 1}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentB, ComponentB>().none).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentB>().none).Count());
+            Assert.AreEqual(2, new HashSet<short> {0, 1}.Intersect(new EcsFilter().NoneOf<ComponentA, ComponentB, ComponentB>().none).Count());
         }
 
         [Test]
